requirements.txt
3==0.0.0
asgiref==3.4.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.26.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.5
django-bootstrap4==15.0.0
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-registration==2.4.1
django-six==1.0.4
django-tinymce==2.5.0
djangorestframework==3.12.4
django-urls==1.1.3
django-utils==0.0.2
django-utils-six==2.0
gunicorn==19.0.0
idna==2.10
importlib-metadata==2.1.1
Pillow==8.2.0
psycopg2==2.8.6
python-decouple==3.1
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.2.0
zipp==3.4.1
